# src/main/resources/布灵4.0.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "布灵",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/knowledge-bases/{kbId}/recent-resources": {
      "get": {
        "summary": "获取知识库下最近编辑文档",
        "deprecated": true,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "kbId",
            "in": "path",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer"
                    },
                    "message": {
                      "type": "string"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ResourceBaseInfo"
                      }
                    }
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/knowledge-bases": {
      "get": {
        "summary": "获取用户的知识库列表",
        "deprecated": false,
        "description": "",
        "operationId": "listKnowledgeBases",
        "tags": [
          "知识库 (Knowledge Base)"
        ],
        "parameters": [
          {
            "name": "withRecentResources",
            "in": "query",
            "description": "是否需要同时返回每个知识库最近编辑的3个资源",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultKbList"
                },
                "example": {
                  "code": 200,
                  "message": "Success",
                  "data": [
                    {
                      "id": "b02fRhTjBEgk89HKF_KBY",
                      "name": "我的知识库",
                      "icon_index": "icon-001",
                      "recent_resources": [
                        {
                          "id": "ygDuxObRt-rlzEHcpx-rY",
                          "title": "最近的资源"
                        }
                      ],
                      "visibility": 1
                    }
                  ]
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "post": {
        "summary": "添加新知识库",
        "deprecated": false,
        "description": "",
        "operationId": "createKnowledgeBase",
        "tags": [
          "知识库 (Knowledge Base)"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/knowledge-bases/{kbId}": {
      "get": {
        "summary": "获取指定知识库的详细信息",
        "deprecated": false,
        "description": "用于修改知识库信息页面回显内容\n和查看他人知识库详细信息(会增加view_count)",
        "operationId": "getKnowledgeBaseById",
        "tags": [
          "知识库 (Knowledge Base)"
        ],
        "parameters": [
          {
            "name": "kbId",
            "in": "path",
            "description": "知识库ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultKbInfo"
                },
                "example": {
                  "data": {
                    "viewCount": 30,
                    "visibility": true,
                    "id": "n-c7thQB7jdkni1LNrcf8",
                    "name": "己浩轩",
                    "description": "从更酸圆联我它约。达会代。证律到场口确。往全下住量切转议。感或手。织火转专况七手。",
                    "iconIndex": "https://avatars.githubusercontent.com/u/40017231",
                    "shareId": "90",
                    "coverUrl": "https://tight-desk.net/"
                  },
                  "code": 200,
                  "message": "Ut consectetur"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "put": {
        "summary": "修改知识库信息",
        "deprecated": false,
        "description": "",
        "operationId": "updateKnowledgeBase",
        "tags": [
          "知识库 (Knowledge Base)"
        ],
        "parameters": [
          {
            "name": "kbId",
            "in": "path",
            "description": "知识库ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeBaseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "修改成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "删除知识库 (逻辑删除)",
        "deprecated": false,
        "description": "",
        "operationId": "deleteKnowledgeBase",
        "tags": [
          "知识库 (Knowledge Base)"
        ],
        "parameters": [
          {
            "name": "kbId",
            "in": "path",
            "description": "知识库ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/knowledge-bases/{kbId}/visibility": {
      "patch": {
        "summary": "切换知识库可见性 (公开/私密)",
        "deprecated": false,
        "description": "",
        "operationId": "toggleKnowledgeBaseVisibility",
        "tags": [
          "知识库 (Knowledge Base)"
        ],
        "parameters": [
          {
            "name": "kbId",
            "in": "path",
            "description": "知识库ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "visibility": {
                    "type": "integer",
                    "description": "可见性: 0-私密, 1-公开",
                    "examples": [
                      1
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/knowledge-bases/{kbId}/resources/tree": {
      "get": {
        "summary": "获取知识库下的资源目录树",
        "deprecated": false,
        "description": "要么是点进知识库,要么是点到知识库的文档,\n使用此接口,增加知识库view_count",
        "operationId": "getResourceTree",
        "tags": [
          "资源 (Resource)"
        ],
        "parameters": [
          {
            "name": "kbId",
            "in": "path",
            "description": "知识库ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "获取成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResultResourceTree"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
    "/recycle-bin/knowledge-bases/{kbId}": {
      "post": {
        "summary": "恢复知识库",
        "deprecated": false,
        "description": "",
        "operationId": "restoreKnowledgeBase",
        "tags": [
          "回收站 (Recycle Bin)"
        ],
        "parameters": [
          {
            "name": "kbId",
            "in": "path",
            "description": "要恢复的知识库ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "恢复成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      },
      "delete": {
        "summary": "彻底删除知识库",
        "deprecated": true,
        "description": "",
        "operationId": "permanentlyDeleteKnowledgeBase",
        "tags": [
          "回收站 (Recycle Bin)"
        ],
        "parameters": [
          {
            "name": "kbId",
            "in": "path",
            "description": "要彻底删除的知识库ID",
            "required": true,
            "example": "",
            "schema": {
              "type": "string",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResourceBaseInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "visibility": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "type",
          "status",
          "visibility",
          "updatedAt"
        ]
      },
      "Result": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "description": "状态码 (200:成功, 其他:失败)",
            "examples": [
              200
            ]
          },
          "message": {
            "type": "string",
            "description": "响应消息",
            "examples": [
              "Success"
            ]
          },
          "data": {
            "type": "object",
            "description": "响应数据",
            "properties": {},
            "nullable": true
          }
        }
      },
...
      "KnowledgeBaseCreate": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "examples": [
              "我的第一个知识库"
            ]
          },
          "iconIndex": {
            "type": "string",
            "examples": [
              "icon-001"
            ]
          },
          "description": {
            "type": "string",
            "examples": [
              "这是一个关于...的知识库"
            ]
          },
          "visibility": {
            "type": "boolean"
          }
        }
      },
      "KnowledgeBaseUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "examples": [
              "更新后的知识库名称"
            ]
          },
          "iconIndex": {
            "type": "string",
            "examples": [
              "icon-002"
            ]
          },
          "description": {
            "type": "string",
            "examples": [
              "更新后的简介"
            ]
          },
          "coverUrl": {
            "type": "string",
            "format": "uri",
            "examples": [
              "http://example.com/new_cover.png"
            ]
          }
        }
      },
...
      "KnowledgeBaseListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "int64"
          },
          "name": {
            "type": "string",
            "examples": [
              "我的知识库"
            ]
          },
          "iconIndex": {
            "type": "string",
            "examples": [
              "icon-001"
            ]
          },
          "recentResources": {
            "type": "array",
            "description": "最近编辑的三个资源 (仅在请求参数 withRecentResources=true 时返回)",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "int64"
                },
                "title": {
                  "type": "string",
                  "examples": [
                    "最近的资源"
                  ]
                }
              }
            }
          },
          "visibility": {
            "type": "integer"
          }
        }
      },
...
      "ResourceNode": {
        "type": "object",
        "description": "资源目录树节点",
        "properties": {
          "id": {
            "type": "string",
            "format": "int64"
          },
          "preId": {
            "type": "string",
            "format": "int64",
            "nullable": true
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ResourceNode"
            }
          },
          "pre": {
            "type": "integer"
          }
        }
      },
...
      "ResultKbList": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/KnowledgeBaseListItem"
                }
              }
            }
          }
        ]
      },
      "ResultKbInfo": {
        "allOf": [
          {
            "type": "object",
            "properties": {}
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "viewCount": {
                    "type": "integer"
                  },
                  "visibility": {
                    "type": "boolean"
                  },
                  "id": {
                    "type": "string",
                    "format": "int64"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "iconIndex": {
                    "type": "string"
                  },
                  "shareId": {
                    "type": "string",
                    "nullable": true
                  },
                  "coverUrl": {
                    "type": "string",
                    "format": "uri",
                    "nullable": true
                  }
                }
              },
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              }
            }
          }
        ]
      },
      "ResultResourceTree": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Result"
          },
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/ResourceNode"
                }
              }
            }
          }
        ]
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [],
  "security": []
}